requirements.txt
transformers==4.30.0
streamlit==1.20.0
torch==2.0.1
tensorflow==2.12.0
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.0
librosa==0.10.0
opencv-python==4.6.0.66
Pillow==9.3.0
requests==2.28.1
tqdm==4.64.1
sentence-transformers==2.2.2
transformers[torch]==4.30.0
transformers[tensorflow]==4.30.0